pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.0.10",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
